% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smf2.R, R/read_smf.R
\name{read_smf}
\alias{read_smf}
\alias{make_note_frame}
\alias{read_smf}
\title{rsmf to note_frame}
\usage{
make_note_frame(smf)

read_smf(file)
}
\arguments{
\item{smf}{rsmf object}

\item{file}{input file path}
}
\description{
rsmf to note_frame

SMF read funcition
}
