% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wav.R
\name{read_wav}
\alias{read_wav}
\title{SMF read funcition}
\usage{
read_wav(file, start_time = 0, end_time = NULL, unit = "sec",
  csound = NULL)
}
\arguments{
\item{file}{input file path}

\item{unit}{unit for the start and end times. now only "sec" is allowed.}

\item{start_times}{start time by sec}

\item{end_times}{end time by sec}
}
\description{
SMF read funcition
}
