% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_smf.R
\name{write_smf}
\alias{write_smf}
\title{SMF write funcition}
\usage{
write_smf(file, rsmf)
}
\arguments{
\item{file}{output file path}

\item{rsmf}{output data}
}
\description{
SMF write funcition
}
